requirements.txt
Flask==2.1.2
psycopg2-binary==2.9.3
google-cloud-bigquery==3.3.1
